--- java/br/com/projetoBase/dto/ClinicaDTO.java
@@ -1,4 +1,0 @@
-package br.com.projetoBase.dto;
-
-public record ClinicaDTO(String nome, String descricao, Long max, Long maxHora) {
-}
