--- java/br/com/projetoBase/dto/ConsultaAgendarDTO.java
@@ -1,6 +1,0 @@
-package br.com.projetoBase.dto;
-
-import br.com.projetoBase.modelo.Consulta;
-
-public record ConsultaAgendarDTO(Long consulta, String paciente) {
-}
