--- java/br/com/projetoBase/dto/DeletarDTO.java
@@ -1,4 +1,0 @@
-package br.com.projetoBase.dto;
-
-public record DeletarDTO(Long id) {
-}
