+++ src/main/java/br/com/projetoBase/dto/UsuarioCadastro.java
@@ -1,16 +1,31 @@
 package br.com.projetoBase.dto;
 
+import java.util.Date;
 
-public record UsuarioCadastro(String user, String password, String nome, String clinica) {
-    public String user() {
+public record UsuarioCadastro(long id,String user, String pass, String nome, String telefone,  Date DataNascimento, String cpf) {
+    
+	public long id() {
+		return id;
+	}
+
+	public String user() {
         return user;
     }
 
     public String pass() {
-        return password;
+        return pass;
     }
 
     public String nome() {
         return nome;
     }
+
+	public String telefone() {
+		return telefone;
+	}
+
+	public Date DataNascimento() {
+		return DataNascimento;
+	}
+    
 }
