+++ src/main/java/br/com/projetoBase/dto/clinica/ClinicaDTO.java
@@ -1,0 +1,6 @@
+package br.com.projetoBase.dto.clinica;
+
+import java.time.LocalTime;
+
+public record ClinicaDTO(String nome, String descricao, Long max, int maxHora, LocalTime inicio, LocalTime fim) {
+}
