+++ src/main/java/br/com/projetoBase/dto/clinica/ConsultasDisponiveisResponseDTO.java
@@ -1,0 +1,6 @@
+package br.com.projetoBase.dto.clinica;
+
+import java.util.List;
+
+public record ConsultasDisponiveisResponseDTO(int limite, List<IntervaloHorarioDTO> intervaloHorarioDTOS) {
+}
