+++ src/main/java/br/com/projetoBase/dto/clinica/IntervaloHorarioDTO.java
@@ -1,0 +1,6 @@
+package br.com.projetoBase.dto.clinica;
+
+import java.time.LocalTime;
+
+public record IntervaloHorarioDTO(LocalTime inicio, LocalTime fim, int quantidadeMax) {
+}
