+++ src/main/java/br/com/projetoBase/dto/clinica/VerificarHorariosRequestDTO.java
@@ -1,0 +1,9 @@
+package br.com.projetoBase.dto.clinica;
+
+import br.com.projetoBase.modelo.Clinica;
+
+import java.time.LocalDate;
+
+
+public record VerificarHorariosRequestDTO(Clinica clinica, LocalDate data) {
+}
