+++ src/main/java/br/com/projetoBase/dto/consulta/AgendarConsultaDTO.java
@@ -1,7 +1,6 @@
-package br.com.projetoBase.dto;
+package br.com.projetoBase.dto.consulta;
 
 import br.com.projetoBase.modelo.Clinica;
-import br.com.projetoBase.modelo.Usuario;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
