+++ src/main/java/br/com/projetoBase/dto/consulta/ConsultaAgendarDTO.java
@@ -1,0 +1,4 @@
+package br.com.projetoBase.dto.consulta;
+
+public record ConsultaAgendarDTO(Long consulta, String paciente) {
+}
