+++ src/main/java/br/com/projetoBase/dto/consulta/DeletarDTO.java
@@ -1,0 +1,4 @@
+package br.com.projetoBase.dto.consulta;
+
+public record DeletarDTO(Long id) {
+}
