+++ src/main/java/br/com/projetoBase/dto/consulta/ListByConsultasDTO.java
@@ -1,4 +1,4 @@
-package br.com.projetoBase.dto;
+package br.com.projetoBase.dto.consulta;
 
 import br.com.projetoBase.modelo.Clinica;
