+++ src/main/java/br/com/projetoBase/dto/consulta/ListByPacienteDTO.java
@@ -1,4 +1,4 @@
-package br.com.projetoBase.dto;
+package br.com.projetoBase.dto.consulta;
 
 public record ListByPacienteDTO(String paciente) {
 }
