+++ src/main/java/br/com/projetoBase/modelo/EntidadeAbstrata.java
