+++ src/main/resources/application.properties
@@ -1,7 +1,7 @@
-spring.datasource.url=jdbc:mysql://localhost:3306/projetobase?createDatabaseIfNotExist=true
+spring.datasource.url=jdbc:mysql://localhost:3306/testeltd?createDatabaseIfNotExist=true
 
 spring.datasource.username=root
-spring.datasource.password=password
+spring.datasource.password=di159753
 
 spring.datasource.tomcat.test-while-idle=true
 spring.datasource.tomcat.validation-query=SELECT 1
